# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Categories Controller",
        "version": "1"
    },
    "paths": {
...
        "/admin/users/{id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "operationId": "92c7b82f0128f5fb2749b6a6122268f9",
                "responses": {
                    "200": {
                        "description": "Fetched User Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminUser"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Remove the specified resource from storage.",
                "operationId": "3eeef084edb4cab4be7b76a63a00d492",
                "responses": {
                    "201": {
                        "description": "The User was successfully deleted"
                    }
                }
            },
            "patch": {
                "tags": [
                    "users"
                ],
                "summary": "Update the specified resource in storage.",
                "operationId": "b31315a11e31c57cb37b41edeb97b4c6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/adminUpdateUser"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminUser"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "$ref": "#/components/parameters/adminAdminIdPathParameter"
                }
            ]
        },
        "/store/orders/{id}/accept": {
            "patch": {
                "tags": [
                    "storeOrders"
                ],
                "operationId": "0c7a5c4324b30ad923fdb1f14704e13f",
                "responses": {
                    "200": {
                        "description": "Order Status Changed Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "$ref": "#/components/parameters/storeOrderIdPathParameter"
                }
            ]
        },
        "/store/orders/{id}/reject": {
            "patch": {
                "tags": [
                    "storeOrders"
                ],
                "operationId": "9b207b7520d9c59d2ccb1ee175e89765",
                "responses": {
                    "200": {
                        "description": "Order Status Changed Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "$ref": "#/components/parameters/storeOrderIdPathParameter"
                }
            ]
        },
...
        "/admin/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get All Users",
                "operationId": "1a238c2c3a338cff776f8344159de6c6",
                "responses": {
                    "200": {
                        "description": "The User was successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/adminUser"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Create a new User.",
                "operationId": "0032ea2ed6cae6dbbb2f30a236bc8c93",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/adminCreateUser"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminUser"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/store/auth/login": {
            "post": {
                "tags": [
                    "storeAuth"
                ],
                "operationId": "7b9803f43b76fb1a9411c7d5c4ad194a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/adminLogin"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "User logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminLoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User name and password or incorrect",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/user/auth/login": {
            "post": {
                "tags": [
                    "userAuth"
                ],
                "operationId": "f8b7cd770f6b03a9de0d2841238b7d8b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/userLogin"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "User logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminLoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "credentials are invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "adminLogin": {
                "properties": {
                    "name": {
                        "type": "string",
                        "default": "admin"
                    },
                    "password": {
                        "type": "string",
                        "default": "admin"
                    }
                },
                "type": "object"
            },
            "adminLoginResponse": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "adminCreateUser": {
                "properties": {
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "dial_code": {
                        "type": "string"
                    },
                    "gender": {
                        "$ref": "#/components/schemas/Gender"
                    },
                    "number": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "adminUpdateUser": {
                "properties": {
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "dial_code": {
                        "type": "string"
                    },
                    "gender": {
                        "$ref": "#/components/schemas/Gender"
                    },
                    "number": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "adminUser": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "dial_code": {
                        "type": "string"
                    },
                    "gender": {
                        "$ref": "#/components/schemas/Gender"
                    },
                    "number": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "LoginResponse": {
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "userLogin": {
                "properties": {
                    "dial_code": {
                        "type": "string",
                        "default": "963"
                    },
                    "number": {
                        "type": "string",
                        "default": "0968259851"
                    },
                    "code": {
                        "type": "string",
                        "default": "123456"
                    }
                },
                "type": "object"
            },
...
            "Gender": {
                "description": "[0 => Male, 1 => Female]",
                "type": "integer",
                "enum": [
                    0,
                    1
                ]
            },
...
        },
        "parameters": {
            "adminAdminIdPathParameter": {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
            "storeOrderIdPathParameter": {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
        },
        "securitySchemes": {
            "ref_to_be_used_below_csrf": {
                "type": "apiKey",
                "name": "X-XSRF-TOKEN",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "ref_to_be_used_below_csrf": []
        }
    ]
}